requirements.txt
pandas>=1.1.3
numpy>=1.19.2
gensim>=3.8.3
requests>=2.24.0
wikipedia>=1.4.0
xlrd>=1.2.0
shortuuid>=1.0.1
spacy>=2.3.2
pyspellchecker>=0.5.5
scikit-learn>=0.23.2
beautifulsoup4>=4.9.3
ftfy>=5.8
progressbar2>=3.53.1
nltk>=3.5
textblob-de>=0.4.3
gdown>=3.12.2